cluster: pull topic_configuration out of types.h
This pulls out topic configuration out of types.h so it can more easily
be depended on. I intend to have cloud_storage depend on just the topic
configuration so the manifest code can move back into v_cloud_storage.
