+++ src/v/cluster/CMakeLists.txt
@@ -109,6 +109,7 @@ v_cc_library(
     metadata_dissemination_service.cc
     types.cc
     remote_topic_properties.cc
+    topic_configuration.cc
     topic_properties.cc
     notification_latch.cc
     topic_table.cc
